Make tags appear on the summary page. Remove code from Root controller
This involves adding a fragment and a page for tags, and also moves the
code for log/shortlog into the fragment rather than forwarding.

Note that we have currently lost all the paging, and also the ability to
(short)?log under a tree. This should come back in another form attached
somewhere else later.

You can see from the fragment shortlog code now that the model interface is
pretty yukky still when you simplify the controller code, but that's fine
right now as it's becoming clearer how to clean stuff in the model up more...