Bootstrap test fixes.
The author tests were failing in a bootstrapped environment as the env
script only looked at the current or the parent directory for
local-lib5. Also fixed the POD tests to look under lib so as to avoid
trawling local-lib5 for unrelated modules.

Fixed the tag and head tests to make use of env.