Drop broken assertions in t/02git_CollectionOfRepositories_FromDirectoryRecursive.t.
The assertions only make sense if Gitalist is operating as a git clone
and is the first repo to be found in its parent directory. This will
be a bit faily for CPAN and would probably make some CPAN testers sad.

This also brings us to 100% test passes on the quenu branch, huzzah!