Fixed tree view for repos with / in name
The "/" in the name of the repo was not escaped. Use uri_for_action to
generate the link to the fragments URI. That implies deleting the
"around" uri_for_action in Gitalist.pm. That seems to be OK because
tests still pass. Also, didn't find any invocation to uri_for_action
with a /fragments path.