Align Vhost code with reality.
Now that we potentially have a different COR per hit (depending on config) the
COR model has changed a lot. Now its COMPONENT will return a G::G::COR-based
class which in turn provides an ACCEPT_CONTEXT method that should return a
further G::G::COR object based on the request. In the simple case, as with the
default config, the same G::G::COR that was built for the COMPONENT is also
returned by ACCEPT_CONTEXT. With the Vhost COR a different G::G::COR will be
returned depending on the related configuration.